Pass down searchedUser from store and render EditUserSettings if it exists
